LiveCoding/config/requirements.txt
django
djangorestframework_simplejwt
django-environ
django-cors-headers
psycopg2-binary